Pull out regular backgrounds to their own file and start fleshing it out a little
