+++ agb/src/display/tiled.rs
@@ -1,20 +1,13 @@
+mod regular_background;
 mod vram_manager;
 
-use core::{
-    alloc::{Allocator, Layout},
-    cell::RefCell,
-    marker::PhantomData,
-    ptr::NonNull,
-};
+use core::{cell::RefCell, marker::PhantomData};
 
-use alloc::vec::Vec;
 pub use vram_manager::{DynamicTile, TileFormat, TileIndex, TileSet, VRamManager};
 
 use crate::agb_alloc::{
     block_allocator::BlockAllocator, bump_allocator::StartEnd, impl_zst_allocator,
 };
-
-use super::Priority;
 
 pub struct BackgroundId(pub(crate) u8);
 
@@ -77,48 +70,6 @@ struct TiledBackgroundModifyables {}
 pub struct TiledBackground<'gba> {
     _phantom: PhantomData<&'gba ()>,
     frame_data: RefCell<TiledBackgroundModifyables>,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[repr(u16)]
-pub enum RegularBackgroundSize {
-    Background32x32 = 0,
-    Background64x32 = 1,
-    Background32x64 = 2,
-    Background64x64 = 3,
-}
-
-impl RegularBackgroundSize {
-    fn size_in_bytes(self) -> usize {
-        match self {
-            RegularBackgroundSize::Background32x32 => 32 * 32 * 2,
-            RegularBackgroundSize::Background64x32 => 64 * 32 * 2,
-            RegularBackgroundSize::Background32x64 => 32 * 64 * 2,
-            RegularBackgroundSize::Background64x64 => 64 * 64 * 2,
-        }
-    }
-
-    fn layout(self) -> Layout {
-        Layout::from_size_align(self.size_in_bytes(), SCREENBLOCK_SIZE).unwrap()
-    }
-}
-
-pub struct RegularBackgroundTiles {
-    priority: Priority,
-    tiles: Vec<Tile>,
-    is_dirty: bool,
-
-    size: RegularBackgroundSize,
-
-    screenblock_ptr: NonNull<u8>,
-}
-
-impl Drop for RegularBackgroundTiles {
-    fn drop(&mut self) {
-        unsafe { ScreenblockAllocator.deallocate(self.screenblock_ptr, self.size.layout()) };
-
-        // TODO: Deallocate the tiles
-    }
 }
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
